backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic[email]==2.8.2
slowapi==0.1.9
python-dotenv==1.0.1
